backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
python-multipart
sqlalchemy
pydantic
passlib
python-jose
numpy
h5py
pillow
torch
torchvision
matplotlib
pydicom
scipy
vtk
email-validator
